requirements.txt
Flask==2.3.2
numpy==1.23.5
opencv-python-headless==4.8.0.74
Pillow==9.5.0
tensorflow-cpu==2.13.0
gunicorn==20.1.0